fix widget import paths after reorganization
