+++ src/widgets/recent-notes/index.ts
@@ -1,6 +1,6 @@
 import { App, TFile, setIcon } from 'obsidian';
-import type { WidgetConfig, WidgetImplementation } from '../interfaces';
-import type WidgetBoardPlugin from '../main';
+import type { WidgetConfig, WidgetImplementation } from '../../interfaces';
+import type WidgetBoardPlugin from '../../main';
 import moment from 'moment';
 
 export interface RecentNotesWidgetSettings {
